requirements.txt
zxcvbn-python
nltk
